+++ resolwe_bio/tests/workflows/test_wgbs.py
@@ -3,6 +3,7 @@ from resolwe.flow.models import Data
 from resolwe.test import tag_process
 
 from resolwe_bio.utils.test import BioProcessTestCase
+from resolwe_bio.utils.filter import filter_comment_lines
 
 
 class WgbsWorkflowTestCase(BioProcessTestCase):
@@ -30,5 +31,23 @@ class WgbsWorkflowTestCase(BioProcessTestCase):
         for data in Data.objects.all():
             self.assertStatus(data, Data.STATUS_DONE)
 
-        wgbs = Data.objects.last()
-        self.assertFile(wgbs, 'hmr', '3A_WT_WGBS_chr2_17k_single.hmr.gz', compression='gzip')
+        hmr = Data.objects.filter(process__slug='hmr').last()
+        self.assertFile(hmr, 'hmr', '3A_WT_WGBS_chr2_17k_single.hmr.gz', compression='gzip')
+        summary = Data.objects.filter(process__slug='alignment-summary').last()
+        self.assertFile(
+            summary,
+            'report',
+            '3A_WT_WGBS_alignment_metrics.txt',
+            file_filter=filter_comment_lines)
+        wgs_metrics = Data.objects.filter(process__slug='wgs-metrics').last()
+        self.assertFile(
+            wgs_metrics,
+            'report',
+            '3A_WT_WGBS_wgs_metrics.txt',
+            file_filter=filter_comment_lines)
+        rrbs_metrics = Data.objects.filter(process__slug='rrbs-metrics').last()
+        self.assertFile(
+            rrbs_metrics,
+            'report',
+            '3A_WT_WGBS_rrbs_summary_metrics.txt',
+            file_filter=filter_comment_lines)